refactor: terminalReducer > 타입, 유틸함수 분리
[#279]
