+++ packages/frontend/src/reducers/terminalReducer/index.ts
@@ -1,0 +1,6 @@
+export {
+  terminalReducer,
+  initialTerminalState,
+  terminalActionTypeMap,
+} from "./terminalReducer";
+export { TerminalActionTypes } from "./type";
